Module srither_core::geom

Structs

CellId
An ID identifying a cell in lattice rectangle.

Move
A difference between two Points.

Point
A two-dimensional lattice point.

Points
An iterator iterating all points in the rectangle.

PointsInColumn
An iterator iterating all points in a column of the rectangle.

PointsInRow
An iterator iterating all points in a row of the rectangle.

Rotation
A 2x2 rotation matrix.

Size
A size of a rectangle.

Table
A table that stores values for each cells.

Traits

Geom
A rectangle area.