Module srither_core::lattice_parser

Parsing a lattice strings.

Structs

Cells
An iterator iterating the cells of the lattice.

HEdges
An iterator iterating the horizontal edges of the lattice.

LatticeParser
A parser parsing a string into lattice.

ParseLatticeError
An error type which is returned from parsing a string into lattice.

VEdges
An iterator iterating the vertical edges of the lattice.