Struct srither_core::lattice_parser::LatticeParser

pub struct LatticeParser<'a> {
    // some fields omitted
}

A parser parsing a string into lattice.

Methods

impl<'a> LatticeParser<'a>

fn from_lines(lines: &'a [Vec<char>]) -> Result<LatticeParser<'a>, ParseLatticeError>

Creates a lattice parser from lines of strings.

fn num_rows(&self) -> usize

Returns the number of the rows.

fn num_cols(&self) -> usize

Returns the number of the columns.

fn v_edges(&self) -> VEdges

Returns an iterator iterating the vertical edges of the lattice.

fn h_edges(&self) -> HEdges

Returns an iterator iterating the horizontal edges of the lattice.

fn cells(&self) -> Cells

Returns an iterator iterating the cells of the lattice.