Struct srither_core::puzzle::Puzzle

pub struct Puzzle {
    // some fields omitted
}

Slither link puzzle data.

Methods

impl Puzzle

fn new(size: Size) -> Puzzle

Creates an empty slither link puzzle.

fn hint(&self, p: Point) -> Hint

Gets a hint at the point.

fn set_hint(&mut self, p: Point, hint: Hint)

Sets a hint at the point.

fn side(&self, p: Point) -> Option<Side>

Gets a side at the point.

fn set_side(&mut self, p: Point, side: Option<Side>)

Set a side at the point.

fn edge_h(&self, p: Point) -> Option<Edge>

Gets a horizontal edge above the point.

fn set_edge_h(&mut self, p: Point, edge: Option<Edge>)

Sets a horizontal edge above the point.

fn edge_v(&self, p: Point) -> Option<Edge>

Gets a vertical edge on the right of the point.

fn set_edge_v(&mut self, p: Point, edge: Option<Edge>)

Sets a vertical edge on the right of the point.